requirements.txt
annotated-types==0.7.0
anyio==4.13.0
babel==2.18.0
beautifulsoup4==4.14.3
certifi==2026.4.22
charset-normalizer==3.4.7
courlan==1.3.2
dateparser==1.4.0
distro==1.9.0
feedparser==6.0.12
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
jusText==3.0.2
lxml==6.1.0
lxml_html_clean==0.4.4
openai==2.36.0
pydantic==2.13.4
pydantic_core==2.46.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.2
PyYAML==6.0.3
regex==2026.5.9
requests==2.33.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tld==0.13.2
tqdm==4.67.3
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.7.0
